refactor: code refactor according to pull request comments
