+++ src/components/timeline/AuthorsTimeline.jsx
@@ -1,34 +1,36 @@
-import React, { useMemo } from "react";
+import React from "react";
 import { Timeline, TimelineItem } from "vertical-timeline-component-for-react";
 import Typography from "@material-ui/core/Typography";
 import PropTypes from "prop-types";
+import theme from "../../config/material.theme";
 
 AuthorsTimeline.propTypes = {
-  biography: PropTypes.object
+  timeLineItems: PropTypes.object
 };
 
 function AuthorsTimeline(props) {
-  const { biography } = props;
-
-  const timeLineItems = useMemo(
-    () =>
-      biography.map((item, index) => {
+  const { timeLineItems } = props;
+  const { palette } = theme;
+  return (
+    <Timeline lineColor={palette.secondary.lightBlue}>
+      {timeLineItems.map((item, index) => {
         return (
           <TimelineItem
             key={index}
             dateText={item.year}
-            dateInnerStyle={{ background: "#49c5b6", color: "#000" }}
+            dateInnerStyle={{
+              background: palette.secondary.main,
+              color: palette.text.secondary
+            }}
           >
             <Typography variant="body2" color="textSecondary" component="p">
               {item.info}
             </Typography>
           </TimelineItem>
         );
-      }),
-    [biography]
+      })}
+    </Timeline>
   );
-
-  return <Timeline lineColor={"#aed1d5"}>{timeLineItems}</Timeline>;
 }
 
 export default AuthorsTimeline;
